+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Address.kt
@@ -1,0 +1,10 @@
+package ru.practicum.android.diploma.domain.models
+
+data class Address(
+    val building: String,
+    val city: String,
+    val description: String,
+    val lat: Double,
+    val lng: Double,
+    val street: String
+)
