+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Area.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.domain.models
+
+data class Area(
+    val id: String,
+    val name: String,
+    val url: String
+)
