+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Contacts.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.domain.models
+
+data class Contacts(
+    val email: String,
+    val name: String,
+    val phones: List<Phone>
+)
