+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Employment.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.domain.models
+
+data class Employment(
+    val id: String,
+    val name: String
+)
