+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Experience.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.domain.models
+
+data class Experience(
+    val id: String,
+    val name: String
+)
