+++ app/src/main/java/ru/practicum/android/diploma/domain/models/KeySkill.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.domain.models
+
+data class KeySkill(
+    val name: String
+)
