+++ app/src/main/java/ru/practicum/android/diploma/domain/models/LogoUrls.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.domain.models
+
+data class LogoUrls(
+    val original: String
+)
