+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Salary.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.domain.models
+
+data class Salary(
+    val currency: String,
+    val from: Int,
+    val gross: Boolean,
+    val to: Any
+)
