+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Schedule.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.domain.models
+
+data class Schedule(
+    val id: String,
+    val name: String
+)
