+++ app/src/main/java/ru/practicum/android/diploma/domain/models/VacancyDescription.kt
@@ -15,72 +15,3 @@ data class VacancyDescription(
     val contacts: Contacts,             //контакты
     val description: String,            //комментарий
 )
-
-data class Address(
-    val building: String,
-    val city: String,
-    val description: String,
-    val lat: Double,
-    val lng: Double,
-    val street: String
-)
-
-data class Area(
-    val id: String,
-    val name: String,
-    val url: String
-)
-
-
-data class Contacts(
-    val email: String,
-    val name: String,
-    val phones: List<Phone>
-)
-
-data class Employer(
-    val alternate_url: String,
-    val blacklisted: Boolean,
-    val id: String,
-    val logo_urls: LogoUrls,
-    val name: String,
-    val trusted: Boolean,
-    val url: String
-)
-
-data class Employment(
-    val id: String,
-    val name: String
-)
-
-data class Experience(
-    val id: String,
-    val name: String
-)
-
-data class KeySkill(
-    val name: String
-)
-
-data class Salary(
-    val currency: String,
-    val from: Int,
-    val gross: Boolean,
-    val to: Any
-)
-
-data class Schedule(
-    val id: String,
-    val name: String
-)
-
-data class Phone(
-    val city: String,
-    val comment: Any,
-    val country: String,
-    val number: String
-)
-
-data class LogoUrls(
-    val original: String
-)
